plugin cmds wip
