+++ Core/alice/commands/quit.hpp
