+++ Core/alice/commands/show.hpp
