+++ Core/alice/commands/write_io.hpp
