--- Core/interfaces/LayoutPluginFactory.cpp
@@ -1,6 +1,0 @@
-#include "LayoutPluginFactory.h"
-
-LayoutPluginFactory::LayoutPluginFactory()
-{
-
-}
