+++ Core/interfaces/Plugin.cpp
@@ -1,0 +1,15 @@
+#include "Plugin.h"
+
+Plugin::Plugin()
+{
+
+}
+
+po::options_description& Plugin::opts() {
+    return mOpts;
+}
+
+po::positional_options_description& Plugin::pods()
+{
+    return mPods;
+}
