+++ Elve/Elve.pro
@@ -34,22 +34,6 @@ HEADERS += \
     MainWindow.h \
     FileLoadAction.h \
     LayoutLoadAction.h \
-    alice/commands/alias.hpp \
-    alice/commands/convert.hpp \
-    alice/commands/current.hpp \
-    alice/commands/help.hpp \
-    alice/commands/print.hpp \
-    alice/commands/ps.hpp \
-    alice/commands/quit.hpp \
-    alice/commands/read_io.hpp \
-    alice/commands/set.hpp \
-    alice/commands/show.hpp \
-    alice/commands/store.hpp \
-    alice/commands/write_io.hpp \
-    alice/alice.hpp \
-    alice/command.hpp \
-    alice/readline.hpp \
-    alice/rules.hpp \
     QConsoleWidget.h \
     Redirect.h \
     uis/Tab.h \
